Added support for network argument
Background: many StarkNet commands take a network argument which allows
a user to select a specific network as the target for the command. Prior
to these changes the StarkNet commands exposed in this box used the
alpha-goerli testnet without the ability to easily change networks.

Added support for a network argument in the starknet:deploy_account and
starknet:deploy commands. A default network can be configured in the
truffle-config.starknet.js file along with the network configurations as
follows:

networks: {
  testnet: {
    network_id: "alpha-goerli",
  },
  default: {
    network: "testnet",
  },
},

The default network is configured as a reference or pointer to any
network that is configured. If a user does not provide a network
argument on the command-line, the default network will be used as
configured. If there is no default network configured then the scripts
will fall back to use alpha-goerli. If the user specifies a network that
is not configured the scripts will again fall back to the default
network.