Added support for StarkNet Devnet
Added support for Shard Labs' StarkNet Devnet. This is a docker image which provides a development testnet similar to Ganache. The Devnet image can be used as a development network for deploying and interacting with StarkNet contracts. It differs to StarkNet Alpha testnet and, so is not a substitute for testing on Alpha testnet. It is not intended as a target network for unit test scripts at this stage.

Additions and modifications:
- added a new StarkNetDevnetDocker class responsible operating on the Devnet docker image.
- added runContainer() method to the TruffleDocker class. The Devnet container, once started continuees to run until stopped by the user. This is in contrast to the other docker images used by the box which execute a command and then stop. This method was needed to run a docker image without supplying a command to be executed by the contatiner on start.
- added configuration to the truffle-config.starknet.js file for the StarkNet Devnet container. The configuration refers to a docker repository and version tag to allow the Devnet image to be located and run, and pulled from the Docker Hub if necessary.
- added a new script (start_devnet.mjs) which is responsible for starting the Devnet container.
- added starknet:start_devnet command to package.json which runs the start_devnet.mjs script.